scratch/requirements.txt
pygithub
httpx
